fixing compiler errors

+++ test/gsection/16g.test.cpp
@@ -260,7 +260,7 @@ void verifyChunkFission( const section::GType< 16 >& chunk ) {
 
     };
 
-    for ( size_t g_i = 0; g_i < chunk.NGN(); ++g_i ) {
+    for ( size_t g_i = 0; g_i < static_cast< std::size_t >( chunk.NGN() ); ++g_i ) {
         CHECK_THAT( expected_flux[0][0][g_i], WithinRel( chunk.flux( 0, 0 )[g_i] ) );
         for ( size_t g_o = 0; g_o < 12; ++g_o ) {
             CHECK_THAT( expected_matrix[0][0][g_i][g_o], WithinRel( chunk.matrix( 0, 0 )[g_i][g_o] ) );
@@ -319,7 +319,7 @@ void verifyChunkInelastic( const section::GType< 16 >& chunk ) {
             }
         }
     };
-    for ( size_t g_i = 0; g_i < chunk.NGN(); ++g_i ) {
+    for ( size_t g_i = 0; g_i < static_cast< std::size_t >( chunk.NGN() ); ++g_i ) {
         CHECK_THAT( expected_flux[0][0][g_i], WithinRel( chunk.flux( 0, 0 )[g_i] ) );
         for ( size_t g_o = 0; g_o < 2; ++g_o ) {
             CHECK_THAT( expected_matrix[0][0][g_i][g_o], WithinRel( chunk.matrix( 0, 0 )[g_i][g_o] ) );
