+++ test/gsection/5/18g.test.cpp
@@ -22,7 +22,7 @@ void verifyChunkPrompt( const section::GType< 5, 18 >& );
 SCENARIO( "section::GType< 5, 18 >" ) {
 
     GIVEN( "valid data for a section::GType< 5, 18 > for prompt fission neutron spectra" ) {
-        
+
         std::string sectionString = chunkPrompt() + validSEND();
 
         WHEN( "the data is provided explicitly" ){
@@ -31,7 +31,7 @@ SCENARIO( "section::GType< 5, 18 >" ) {
             double awr = 0.0;
             int lr = 0;
             double temp = 293.6;
-            std::vector< double > chi = { 1.96730e-14, 1.15287e-11, 7.10093e-11, 7.358034e-9, 9.678216e-7, 9.976067e-5, 
+            std::vector< double > chi = { 1.96730e-14, 1.15287e-11, 7.10093e-11, 7.358034e-9, 9.678216e-7, 9.976067e-5,
                   8.080721e-4, 4.412081e-3, 3.119482e-3, 2.172044e-2, 4.250824e-2, 1.727228e-1,
                   6.932789e-2, 1.778175e-1, 2.204227e-1, 2.772107e-1, 8.677879e-3, 1.124378e-3,
                   1.884936e-5, 8.232639e-6 };
@@ -84,11 +84,11 @@ SCENARIO( "section::GType< 5, 18 >" ) {
 
 std::string chunkPrompt() {
     return
-   " 9.223500+4 0.000000+0          1          1          0         209228 5 18     \n"    
-   " 2.936000+2 0.000000+0         20          1         20         209228 5 18     \n"    
-   " 1.96730-14 1.15287-11 7.10093-11 7.358034-9 9.678216-7 9.976067-59228 5 18     \n"    
-   " 8.080721-4 4.412081-3 3.119482-3 2.172044-2 4.250824-2 1.727228-19228 5 18     \n"    
-   " 6.932789-2 1.778175-1 2.204227-1 2.772107-1 8.677879-3 1.124378-39228 5 18     \n"    
+   " 9.223500+4 0.000000+0          1          1          0         209228 5 18     \n"
+   " 2.936000+2 0.000000+0         20          1         20         209228 5 18     \n"
+   " 1.96730-14 1.15287-11 7.10093-11 7.358034-9 9.678216-7 9.976067-59228 5 18     \n"
+   " 8.080721-4 4.412081-3 3.119482-3 2.172044-2 4.250824-2 1.727228-19228 5 18     \n"
+   " 6.932789-2 1.778175-1 2.204227-1 2.772107-1 8.677879-3 1.124378-39228 5 18     \n"
    " 1.884936-5 8.232639-6                                            9228 5 18     \n";
 }
 
@@ -111,11 +111,11 @@ void verifyChunkPrompt( const section::GType< 5, 18 >& chunk ) {
     CHECK( 0 == chunk.breakUp() );
     CHECK_THAT( 293.6, WithinRel( chunk.TEMP() ) );
     CHECK_THAT( 293.6, WithinRel( chunk.temperature() ) );
-    std::vector< double > expected_chi = { 1.96730e-14, 1.15287e-11, 7.10093e-11, 7.358034e-9, 9.678216e-7, 9.976067e-5, 
+    std::vector< double > expected_chi = { 1.96730e-14, 1.15287e-11, 7.10093e-11, 7.358034e-9, 9.678216e-7, 9.976067e-5,
                                            8.080721e-4, 4.412081e-3, 3.119482e-3, 2.172044e-2, 4.250824e-2, 1.727228e-1,
                                            6.932789e-2, 1.778175e-1, 2.204227e-1, 2.772107e-1, 8.677879e-3, 1.124378e-3,
                                            1.884936e-5, 8.232639e-6 };
-    for ( size_t g = 0; g < chunk.NGN(); ++g ) {
+    for ( size_t g = 0; g < static_cast< std::size_t >( chunk.NGN() ); ++g ) {
         CHECK_THAT( expected_chi[g], WithinRel( chunk.chi()[g] ) );
     }
